Moved prices to serverjs file
