+++ src/cardpage.js
@@ -2,6 +2,8 @@ import React, { Component } from 'react';
 import ReactDOM from 'react-dom';
 import { Card, CardHeader, Row, Col, CardBody, CardText } from 'reactstrap';
 
+import ImageFallback from './components/ImageFallback';
+
 class CardPage extends Component {
   constructor(props) {
     super(props);
@@ -9,10 +11,26 @@ class CardPage extends Component {
 
   render() {
     const {card, data} = this.props;
+    console.log(card);
     return (
         <Card>
-            <CardHeader>{card.name}</CardHeader>
-            <CardBody><pre>{JSON.stringify(data, null, 2)}</pre> </CardBody>
+            <CardHeader><h4>{card.name}</h4></CardHeader>
+            <CardBody>
+              <Row>
+                <Col xs="12" sm="4">
+                  <ImageFallback
+                    className="w-100"
+                    src={card.image_normal}
+                    fallbackSrc="/content/default_card.png"
+                    alt={card.name}
+                  />
+                  <div className="price-area">
+                  </div>
+                </Col>
+              <Col xs="12" sm="8">
+              </Col>
+            </Row>
+            </CardBody>
       </Card>
     );
   }
